feat: improve function interface by strategies
