+++ src/index.ts
@@ -1,17 +1,57 @@
 import * as lambda from 'aws-cdk-lib/aws-lambda';
 import { Construct } from 'constructs';
 
-import lambdaEnvVars, {
-  ITokenValidationStrategy,
-  ITokenValidationStrategyAjvJsonSchemaValidator,
-  TokenValidationStrategy,
-  ITokenVerificationStrategy,
-  ITokenVerificationStrategyArgument,
-  ITokenVerificationStrategyJwksFromUriByKid,
-  TokenVerificationStrategy,
-  TokenAuthorizerOptions,
-  TokenAuthorizerFunctionOptions,
-} from './options';
+// note: unfortunately C# jsii compilation does not allow a common interface with strategyName neither a union type
+
+export interface ITokenValidationStrategyAjvJsonSchemaValidator {
+  readonly strategyName: 'schema';
+  readonly schema: string;
+}
+
+export interface ITokenVerificationStrategyArgument {
+  readonly strategyName: 'argument';
+  readonly secret: string;
+}
+
+export interface ITokenVerificationStrategyJwksFromUriByKid {
+  readonly strategyName: 'jwksFromUriByKid';
+  readonly uri: string;
+  readonly kid: string;
+}
+
+export interface ITokenAuthorizerOptions {
+  readonly verificationStrategy: ITokenVerificationStrategyArgument | ITokenVerificationStrategyJwksFromUriByKid;
+  readonly payloadValidationStrategy?: ITokenValidationStrategyAjvJsonSchemaValidator;
+}
+
+export interface TokenAuthorizerFunctionOptions extends lambda.FunctionOptions {
+  readonly tokenAuthorizerOptions: ITokenAuthorizerOptions;
+}
+
+enum FnEnvVars {
+  JwksUri = 'TOKEN_AUTHORIZER_JWKS_URI',
+  JwksKid = 'TOKEN_AUTHORIZER_JWKS_KID',
+  Secret = 'TOKEN_AUTHORIZER_JWT_VERIFICATION_SECRET',
+  Schema = 'TOKEN_AUTHORIZER_JWT_VALIDATOR_SCHEMA_JSON'
+}
+
+const lambdaEnvVars = (options: TokenAuthorizerFunctionOptions): any => {
+  let e = ((options!.environment ?? {}) as any);
+  switch (options.tokenAuthorizerOptions!.payloadValidationStrategy?.strategyName || '') {
+    case 'schema':
+      e[FnEnvVars.Schema] = options.tokenAuthorizerOptions!.payloadValidationStrategy!.schema;
+  }
+  switch (options.tokenAuthorizerOptions.verificationStrategy.strategyName) {
+    case 'argument':
+      e[FnEnvVars.Secret] = options.tokenAuthorizerOptions.verificationStrategy.secret;
+  }
+  switch (options.tokenAuthorizerOptions.verificationStrategy.strategyName) {
+    case 'jwksFromUriByKid':
+      e[FnEnvVars.JwksUri] = options.tokenAuthorizerOptions.verificationStrategy.uri;
+      e[FnEnvVars.JwksKid] = options.tokenAuthorizerOptions.verificationStrategy.kid;
+  }
+  return e;
+};
 
 export class TokenAuthorizerJwtFunction extends lambda.Function {
   public constructor(scope: Construct, id: string, props: TokenAuthorizerFunctionOptions) {
@@ -28,15 +68,3 @@ export class TokenAuthorizerJwtFunction extends lambda.Function {
     });
   }
 }
-
-export {
-  ITokenValidationStrategy,
-  ITokenValidationStrategyAjvJsonSchemaValidator,
-  TokenValidationStrategy,
-  ITokenVerificationStrategy,
-  ITokenVerificationStrategyArgument,
-  ITokenVerificationStrategyJwksFromUriByKid,
-  TokenVerificationStrategy,
-  TokenAuthorizerOptions,
-  TokenAuthorizerFunctionOptions,
-};
